refact: form structure
