+++ src/app/hira/medication-overall/page.tsx
@@ -2,7 +2,7 @@
 import { useState } from "react";
 import { FormDataType, MultiFactorInfo, FormError } from "@/app/types/hira";
 import Button from "@/app/components/Button";
-import HiraForm from "@/app/components/HiraForm";
+import HiraForm from "@/app/components/form/HiraForm";
 import { validateField, validateForm } from "@/app/utils/nhis-validation";
 
 export default function NhisPage() {
